chore: refactor cache interceptors
