+++ .changeset/rude-frogs-run.md
@@ -4,8 +4,9 @@
 
 Update public API:
 
-- AjaxClient is now Ajax
-- AjaxClientFetchError is now AjaxFetchError
-- request and requestJson methods of Ajax class are renamed as fetch and fetchJson respectively
-- getCookie and validateOptions is not part of the public API any more
+- `AjaxClient` is now `Ajax`
+- `AjaxClientFetchError` is now `AjaxFetchError`
+- `request` and `requestJson` methods of `Ajax` class are renamed as `fetch` and `fetchJson` respectively
+- `getCookie` and `validateOptions` is not part of the public API any more
 - Removed the global `ajax` instance and `setAjax`
+- Exporting `createCacheInterceptors` instead of `cacheRequestInterceptorFactory` and `cacheResponseInterceptorFactory`
