+++ packages/ajax/src/interceptors/acceptLanguageHeader.js
