+++ packages/ajax/src/interceptors/index.js
@@ -1,4 +1,3 @@
-export { acceptLanguageRequestInterceptor } from './acceptLanguageRequestInterceptor.js';
-export { createXsrfRequestInterceptor } from './createXsrfRequestInterceptor.js';
-export { createCacheRequestInterceptor } from './createCacheRequestInterceptor.js';
-export { createCacheResponseInterceptor } from './createCacheResponseInterceptor.js';
+export { acceptLanguageRequestInterceptor } from './acceptLanguageHeader.js';
+export { createXsrfRequestInterceptor } from './xsrfHeader.js';
+export { createCacheInterceptors } from './cacheInterceptors.js';
