+++ packages/ajax/src/interceptors/xsrfHeader.js
